+++ crates/resolver-tests/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name = "resolver-tests"
+version = "0.1.0"
+authors = ["Alex Crichton <alex@alexcrichton.com>"]
+edition = "2018"
+
+[dependencies]
+cargo = { path = "../.." }
+proptest = "0.9.1"
+lazy_static = "1.3.0"
+varisat = "0.2.1"
+atty = "0.2.11"
